Refactor actor and graph interfaces by adding detailed comments for clarity; introduce Routable and Node interfaces in node.go for enhanced routing functionality.
